backend/pyproject.toml
[project]
name = "kidlearn-backend"
version = "0.1.0"
description = "Backend API for KidLearn education platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
